Offer controller implemented and integrated with offerscrapper

+++ src/main/java/zpi/flattery/webscrapper/strategy/ScrapStrategy.java
@@ -1,16 +1,15 @@
 package zpi.flattery.webscrapper.strategy;
 
-import zpi.flattery.model.Offer;
-import zpi.flattery.model.OfferType;
-import zpi.flattery.model.RoomType;
+import zpi.flattery.models.Offer;
+import zpi.flattery.models.enums.OfferType;
+import zpi.flattery.models.enums.RoomType;
 
-import java.util.ArrayList;
 import java.util.List;
 
 public interface ScrapStrategy {
     void scrap();
 
-    void setScrapParameters(RoomType roomType, OfferType offerType, String query, Integer minPrice, Integer maxPrice, String place, Integer radius, Integer daysOld);
+    void setScrapParameters(RoomType roomType, OfferType offerType, String query, double minPrice, double maxPrice, String place, Integer radius, Integer daysOld);
 
     List<Offer> getScrappedOffers();
 }
